Extend encoder output with stub pv
We extend the h vectors after the encoder just before the decoder.
The shapes of 8 param matrices change, but those or handled by a
single variable.  We also need to feed in the extra pv values into the
graph.  Added a stub to place 500 zeros there.  The tiny training
example runs correctly, so this appears to be working.

Incidentally, found two nematus bugs.  One in TextIterator mistakenly
allows examples with source or target length larger than maxlen
tokens.  But this bug is hidden because code elsewhere reruns this
maxlen filtering.  There's a fencepost error in that code that
mistakenly disallows sentences with maxlen tokens.  maxlen - 1 is the
current real max used.